fix: impl_core symlink
